Instructions ArrayList<Method> Add to AntBrain
The getInstruction method should return a set of instructions for the
ant to be able to go through and the array list and call the methods it
needs to in the world. Example of this happening is in the first case in
the switch of getting the direction of the ant and adding that method
into the arrayList which holds all the methods which should be called.